Add basic support for GPTL timing
Add build options for GPTL and AUTOPROFILING.
Both options are OFF by default.

If GPTL is 1 or "on", it will build with support for GPTL timers.
If AUTOPROFILING is 1 or "on", it will build with autoprofiling enabled.

sw_driver.f90 had to be renamed with .F90 so that preprocessing
for "#ifdef GPTL_ENABLED" would be done automatically by the compiler.